Add timemod.py
Add timemod.py which is a command line tool that lets the user batch rename files if the filenames match the yyyymmdd_hhMMSS_string.jpg format. This may come in handy when the exif date and time setting was differently set on two or more cameras and when using images from these cameras, the images appear in the wrong (non-chronological)  order.